projet_stage_Django/requirements.txt
asgiref==3.5.2
attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.7.1
fonttools==4.33.3
future==0.18.2
idna==3.3
isodate==0.6.1
kiwisolver==1.4.2
lxml==4.9.1
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.2
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-file==1.5.1
requests-toolbelt==0.9.1
ScalaFunctional==0.6.0
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
zeep==4.1.0
